data_platform/dashboard/requirements.txt
streamlit>=1.34,<2.0
pandas>=2.0,<3.0
plotly>=5.20,<6.0
requests>=2.31,<3.0
deltalake>=0.17,<1.0
PyYAML>=6.0,<7.0